Update architecture v3
