+++ src/domain/schemas/users.py
@@ -16,6 +16,11 @@ class UserRegister(BaseModel):
             raise ValueError('Номер телефона должен начинаться с "+" и содержать от 5 до 15 цифр')
         return value
 
+
 class UserAuth(BaseModel):
     email: EmailStr = Field(..., description="Электронная почта")
     password: str = Field(..., min_length=5, max_length=50, description="Пароль, от 5 до 50 знаков")
+
+
+class UserChangeRole(BaseModel):
+    email: EmailStr = Field(..., description="Электронная почта")
