+++ src/infrastructure/db/dao/users.py
@@ -1,0 +1,6 @@
+from src.infrastructure.db.dao.base import BaseDAO
+from src.infrastructure.db.database import User
+
+
+class UsersDAO(BaseDAO):
+    model = User
