+++ src/infrastructure/db/database.py
@@ -4,7 +4,7 @@ from typing import Annotated
 from sqlalchemy import func
 from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker, AsyncAttrs
 from sqlalchemy.orm import DeclarativeBase, declared_attr, Mapped, mapped_column
-
+from sqlalchemy import text
 from src.configurator.config import get_db_url
 
 DATABASE_URL = get_db_url()
@@ -28,3 +28,21 @@ class Base(AsyncAttrs, DeclarativeBase):
 
     created_at: Mapped[created_at]
     updated_at: Mapped[updated_at]
+
+
+class User(Base):
+    id: Mapped[int_pk]
+    phone_number: Mapped[str_uniq]
+    first_name: Mapped[str]
+    last_name: Mapped[str]
+    email: Mapped[str_uniq]
+    password: Mapped[str]
+
+    is_user: Mapped[bool] = mapped_column(default=True, server_default=text('true'), nullable=False)
+    is_admin: Mapped[bool] = mapped_column(default=False, server_default=text('false'), nullable=False)
+    is_super_admin: Mapped[bool] = mapped_column(default=False, server_default=text('false'), nullable=False)
+
+    extend_existing = True
+
+    def __repr__(self):
+        return f"{self.__class__.__name__}(id={self.id})"
