+++ src/infrastructure/db/migration/env.py
@@ -1,17 +1,5 @@
 import asyncio
 from logging.config import fileConfig
-
-from sqlalchemy import pool
-from sqlalchemy.engine import Connection
-from sqlalchemy.ext.asyncio import async_engine_from_config
-
-from alembic import context
-
-import sys
-from os.path import dirname, abspath
-
-import asyncio
-from logging.config import fileConfig
 from sqlalchemy import pool
 from sqlalchemy.engine import Connection
 from sqlalchemy.ext.asyncio import async_engine_from_config
@@ -22,8 +10,7 @@ from os.path import dirname, abspath
 
 sys.path.insert(0, dirname(dirname(abspath(__file__))))
 
-from src.adapters.db.database import DATABASE_URL, Base
-from src.domain.models.users.users import User
+from src.infrastructure.db.database import DATABASE_URL, Base
 
 config = context.config
 config.set_main_option("sqlalchemy.url", DATABASE_URL)
